Be even more permissive in our date parsing logic
Not that we aren't already insanely permissive in parsing just about any
random noise that could _possibly_ be construed as xml and turn it into
a dive, this makes us even laxer.

If somebody wants to have a <date> tag with both date and time, why the
heck not? It's fine.  And if it has just the date, that's fine too.  And
the date can be in any of several formats.  We really don't care, the
more permissive, the better.

We strive to always write beautiful xml, but let's face it, not
everybody else does.  If we can turn random line noise into a dive, we
should do so.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>