parse-xml: allow XML nodes with empty tag names
They happen for CDATA content, where libxml2 turns the CDATA fields into
a child of the parent entry, but without a name.

Now, of course, any sane person would just want to use the CDATA as the
string value of the parent itself, but libxml2 probably does this
insanity for a reason.  And the reason is probably that some misguided
people want to *write* XML using libxml2, and then the stupid child node
actually acts as a "now I want you to write this data as CDATA".

Whatever the reason, let's just ignore it.  We will just traverse such a
nameless child and be happy, and we'll give the nameless child the name
of the parent.  Our XML node matching logic will then never see this
insane nameless child at all, and doesn't have to care.

Our whole XML parsing rule-of-thumb is to take the whole "be strict in
what you output, but generous in what you accept" to its logical
conclusion.  Because we will literally accept almost anything, in any
format.  You can mix tags or attributes wildly, and youc an use CDATA or
not as you see fit.  We just don't care.

We're the honeybadger of the divelog world.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>