Display air pressure in the Info notebook
I can't find it in myself to care enough to display this in other units.
mm/Hg anyone? I didn't think so.

We still can't edit this value, but at least if w are able to read it from
the dive computer we also show it to the user.

See #19

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>