Add the begin/end markers to events with the proper flags
When we create the event names, the name itself does not include the
information about whether the event is the beginning or end of some
state, so we end up having things like events named "deco" and then in
the event flags it says whether this is the *beginning* of deco, or the
end.

And when we show the event, we only used to show the name.  This patch
makes us show whether it's the begin or end event for events that have
those flags.  So now you see "deco begin" and "deco end" instead of just
two events both called "deco".

It would perhaps be nice if we somehow showed the range between the
events too, and paired them up visually some way, but that's a separate
and much more difficult thing to do.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>