Updated bg_BG translation for 3.0
- Added translations for pretty much everything new since the last iteration

- Removed most usages of tags where both words begin with upper case
letters. For example, "Some Tag" may look OK in English, but not so good
in other languages. It's much better to have it as "Some tag" in Bulgarian.

- Reverted some of the made-up abbreviations (such as the one for SAC)
to the default English ones (i.e. just SAC). I'm not exactly sure those
made any sense at all, so lets have them in English, as a little precaution.

Hopefully enough for the 10-15 (in total) Bulgarian diving nerds
out there. :)

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>