Fix rounding of GPS coordinates
The whole "+ 0.5" to round to integers only works for positive values,
and GPS coordinates are signed.

So use the proper "round to int" function (rint()), which does this correctly.

Also, remove the redundant check against the master gps values: we
already checked that if we do have a master dive, the gps values must
match the currently edited dive, so comparing against the master is
entirely redundant.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>