Set maxdepth correctly for dives with no samples
This showed up when suddenly some of the test dives no longer got merged
when loaded twice. As I moved maxdepth up into the dive structure and
added the code to fixup the data from what is in the divecomputer I missed
the part where the function is exited early if there are no samples. This
patch corrects that oversight.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>