Add Polish translation
Some acronyms like MOD, SAC etc are left 'as is' as they
are widely known and used in polish diving community
so thinking up a polish version would be plain confusing.

Most strings are translated but there are some that do not show
in my poedit (like "select events" / "gaschange") or I was
not sure about.

Signed-off-by: Krzysztof Arentowicz <k.arentowicz@gmail.com>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>