Modify table print for using pango_layout_get_extents()
1.-  On draw_table() we now set the layout, make the scaling *0.5, and the
     make traslations row by row.
2.-  Wipe out print_table_header and print_table as they now don't make
     anything but call show_table which now we can call directly from
     draw_table.
3.-  Modify show_table to get account of the height of the higher column
     and return it.
4.-  The frame is now plot with an auxiliar func. directly from draw_table.

With this changes we have a variable height pattern which lets us have
complete the more sensible info  (date, time, location, ...) without
wrapping it to fit in the predefined width of the layout.

The printout is tighter than old one so we can print 35 dives in a page,
leaving enough room at the bottom to fit some extra-large (three lines or
more) dives' info.  Have increased *1.2 the size of the fonts as they got
too little with the change in DPI.  Wrap the header strings to limit the
impact of different string sizes in the different supported languages.

Signed-off-by: Salvador Cuñat <salvador.cunat@gmail.com>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>